Abstract lacp search info to own component | update function names and url params
